feat: complete service implementations and integrate RabbitMQ messaging
- Architecture: Completed Clean Architecture implementation for Warehouse, Supplier, Production, and Loyalty services.
- Messaging: Replaced MockEventPublisher with a real RabbitMqEventPublisher using the Aspire RabbitMQ Client across all services.
- Production: Refactored to a job-based state machine (BakingJob). Implemented OrderPlacedConsumer to listen for incoming orders via RabbitMQ. Integrated Redis for real-time job status tracking.
- Loyalty: Implemented customer tier (Bronze, Silver, Gold) and points management with transaction history.
- Supplier: Shifted to a SupplierTransport model to mock incoming ingredient deliveries.
- Warehouse: Standardized quantity properties to decimal across entities and DTOs.
- Frontend: Added modern Blazor pages (Warehouse, Production Tracker, Loyalty Rewards) utilizing Tailwind CSS.
- AppHost: Fixed service timeouts by correctly linking the rabbitmq container to all backend services.

+++ Warehouse/Warehouse.Domain/Events/WarehouseItemCreatedEvent.cs
@@ -1,3 +1,3 @@
 namespace Warehouse.Domain.Events;
 
-public record WarehouseItemCreatedEvent(Guid Id, string Name, double InitialQuantity, string Unit);
+public record WarehouseItemCreatedEvent(Guid Id, string Name, decimal InitialQuantity, string Unit);
