+++ Warehouse/Warehouse.Domain/Exceptions/InsufficientStockException.cs
@@ -1,4 +1,4 @@
 namespace Warehouse.Domain.Exceptions;
 
-public class InsufficientStockException(string name, double requested, double available, string unit) 
+public class InsufficientStockException(string name, decimal requested, decimal available, string unit) 
     : DomainException($"Insufficient stock for '{name}'. Requested: {requested}{unit}, Available: {available}{unit}.");
